Move data objects to its own namespace
