+++ src/Data/Traits/PropertyCityTrait.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Eclipxe\SepomexPhp\Traits;
+namespace Eclipxe\SepomexPhp\Data\Traits;
 
-use Eclipxe\SepomexPhp\City;
+use Eclipxe\SepomexPhp\Data\City;
 use LogicException;
 
 trait PropertyCityTrait
